add .exe file for 2 scripts and arrange folder
