--- Project/_ContactList.csv
@@ -1,3 +1,0 @@
-Name,Phone,Email,DOB,Category
-Test2,222,two.mail,4/8/99,Family
-Test3,333,three,6/8/96,Friends
