--- Project/_ContactList_BST.csv
@@ -1,7 +1,0 @@
-Name,Phone,Email,DOB,Category
-tom,123,tom.mail,,
-data,456,data.mail,,Friends
-nan,951,,5/9/21,
-John,789,john.mail,8/9/13,Friends
-mike,456,,,Friends
-ant,7854,ant.mail,855,Work
